requirements.txt
ttkbootstrap>=0.30.0
opencv-python-headless>=4.8.0
Pillow>=9.0.0
